Polyfill: Refactor since/until to match spec
Introduces the DRY operations ES.DifferenceTemporal___ in order to match
recent changes to the spec text.
