+++ polyfill/lib/plaindatetime.mjs
@@ -1,6 +1,6 @@
 import { ES } from './ecmascript.mjs';
 import { DateTimeFormat } from './intl.mjs';
-import { GetIntrinsic, MakeIntrinsicClass } from './intrinsicclass.mjs';
+import { MakeIntrinsicClass } from './intrinsicclass.mjs';
 
 import {
   ISO_YEAR,
@@ -355,164 +355,11 @@ export class PlainDateTime {
   }
   until(other, options = undefined) {
     if (!ES.IsTemporalDateTime(this)) throw new TypeError('invalid receiver');
-    other = ES.ToTemporalDateTime(other);
-    const calendar = GetSlot(this, CALENDAR);
-    const otherCalendar = GetSlot(other, CALENDAR);
-    const calendarId = ES.ToString(calendar);
-    const otherCalendarId = ES.ToString(otherCalendar);
-    if (calendarId !== otherCalendarId) {
-      throw new RangeError(`cannot compute difference between dates of ${calendarId} and ${otherCalendarId} calendars`);
-    }
-    options = ES.GetOptionsObject(options);
-    const smallestUnit = ES.ToSmallestTemporalUnit(options, 'nanosecond');
-    const defaultLargestUnit = ES.LargerOfTwoTemporalUnits('day', smallestUnit);
-    const largestUnit = ES.ToLargestTemporalUnit(options, 'auto', [], defaultLargestUnit);
-    ES.ValidateTemporalUnitRange(largestUnit, smallestUnit);
-    const roundingMode = ES.ToTemporalRoundingMode(options, 'trunc');
-    const roundingIncrement = ES.ToTemporalDateTimeRoundingIncrement(options, smallestUnit);
-
-    let { years, months, weeks, days, hours, minutes, seconds, milliseconds, microseconds, nanoseconds } =
-      ES.DifferenceISODateTime(
-        GetSlot(this, ISO_YEAR),
-        GetSlot(this, ISO_MONTH),
-        GetSlot(this, ISO_DAY),
-        GetSlot(this, ISO_HOUR),
-        GetSlot(this, ISO_MINUTE),
-        GetSlot(this, ISO_SECOND),
-        GetSlot(this, ISO_MILLISECOND),
-        GetSlot(this, ISO_MICROSECOND),
-        GetSlot(this, ISO_NANOSECOND),
-        GetSlot(other, ISO_YEAR),
-        GetSlot(other, ISO_MONTH),
-        GetSlot(other, ISO_DAY),
-        GetSlot(other, ISO_HOUR),
-        GetSlot(other, ISO_MINUTE),
-        GetSlot(other, ISO_SECOND),
-        GetSlot(other, ISO_MILLISECOND),
-        GetSlot(other, ISO_MICROSECOND),
-        GetSlot(other, ISO_NANOSECOND),
-        calendar,
-        largestUnit,
-        options
-      );
-
-    const relativeTo = ES.TemporalDateTimeToDate(this);
-    ({ years, months, weeks, days, hours, minutes, seconds, milliseconds, microseconds, nanoseconds } =
-      ES.RoundDuration(
-        years,
-        months,
-        weeks,
-        days,
-        hours,
-        minutes,
-        seconds,
-        milliseconds,
-        microseconds,
-        nanoseconds,
-        roundingIncrement,
-        smallestUnit,
-        roundingMode,
-        relativeTo
-      ));
-    ({ days, hours, minutes, seconds, milliseconds, microseconds, nanoseconds } = ES.BalanceDuration(
-      days,
-      hours,
-      minutes,
-      seconds,
-      milliseconds,
-      microseconds,
-      nanoseconds,
-      largestUnit
-    ));
-
-    const Duration = GetIntrinsic('%Temporal.Duration%');
-    return new Duration(years, months, weeks, days, hours, minutes, seconds, milliseconds, microseconds, nanoseconds);
+    return ES.DifferenceTemporalPlainDateTime('until', this, other, options);
   }
   since(other, options = undefined) {
     if (!ES.IsTemporalDateTime(this)) throw new TypeError('invalid receiver');
-    other = ES.ToTemporalDateTime(other);
-    const calendar = GetSlot(this, CALENDAR);
-    const otherCalendar = GetSlot(other, CALENDAR);
-    const calendarId = ES.ToString(calendar);
-    const otherCalendarId = ES.ToString(otherCalendar);
-    if (calendarId !== otherCalendarId) {
-      throw new RangeError(`cannot compute difference between dates of ${calendarId} and ${otherCalendarId} calendars`);
-    }
-    options = ES.GetOptionsObject(options);
-    const smallestUnit = ES.ToSmallestTemporalUnit(options, 'nanosecond');
-    const defaultLargestUnit = ES.LargerOfTwoTemporalUnits('day', smallestUnit);
-    const largestUnit = ES.ToLargestTemporalUnit(options, 'auto', [], defaultLargestUnit);
-    ES.ValidateTemporalUnitRange(largestUnit, smallestUnit);
-    const roundingMode = ES.ToTemporalRoundingMode(options, 'trunc');
-    const roundingIncrement = ES.ToTemporalDateTimeRoundingIncrement(options, smallestUnit);
-
-    let { years, months, weeks, days, hours, minutes, seconds, milliseconds, microseconds, nanoseconds } =
-      ES.DifferenceISODateTime(
-        GetSlot(this, ISO_YEAR),
-        GetSlot(this, ISO_MONTH),
-        GetSlot(this, ISO_DAY),
-        GetSlot(this, ISO_HOUR),
-        GetSlot(this, ISO_MINUTE),
-        GetSlot(this, ISO_SECOND),
-        GetSlot(this, ISO_MILLISECOND),
-        GetSlot(this, ISO_MICROSECOND),
-        GetSlot(this, ISO_NANOSECOND),
-        GetSlot(other, ISO_YEAR),
-        GetSlot(other, ISO_MONTH),
-        GetSlot(other, ISO_DAY),
-        GetSlot(other, ISO_HOUR),
-        GetSlot(other, ISO_MINUTE),
-        GetSlot(other, ISO_SECOND),
-        GetSlot(other, ISO_MILLISECOND),
-        GetSlot(other, ISO_MICROSECOND),
-        GetSlot(other, ISO_NANOSECOND),
-        calendar,
-        largestUnit,
-        options
-      );
-
-    const relativeTo = ES.TemporalDateTimeToDate(this);
-    ({ years, months, weeks, days, hours, minutes, seconds, milliseconds, microseconds, nanoseconds } =
-      ES.RoundDuration(
-        years,
-        months,
-        weeks,
-        days,
-        hours,
-        minutes,
-        seconds,
-        milliseconds,
-        microseconds,
-        nanoseconds,
-        roundingIncrement,
-        smallestUnit,
-        ES.NegateTemporalRoundingMode(roundingMode),
-        relativeTo
-      ));
-    ({ days, hours, minutes, seconds, milliseconds, microseconds, nanoseconds } = ES.BalanceDuration(
-      days,
-      hours,
-      minutes,
-      seconds,
-      milliseconds,
-      microseconds,
-      nanoseconds,
-      largestUnit
-    ));
-
-    const Duration = GetIntrinsic('%Temporal.Duration%');
-    return new Duration(
-      -years,
-      -months,
-      -weeks,
-      -days,
-      -hours,
-      -minutes,
-      -seconds,
-      -milliseconds,
-      -microseconds,
-      -nanoseconds
-    );
+    return ES.DifferenceTemporalPlainDateTime('since', this, other, options);
   }
   round(roundTo) {
     if (!ES.IsTemporalDateTime(this)) throw new TypeError('invalid receiver');
