+++ polyfill/lib/zoneddatetime.mjs
@@ -376,203 +376,11 @@ export class ZonedDateTime {
   }
   until(other, options = undefined) {
     if (!ES.IsTemporalZonedDateTime(this)) throw new TypeError('invalid receiver');
-    other = ES.ToTemporalZonedDateTime(other);
-    const calendar = GetSlot(this, CALENDAR);
-    const otherCalendar = GetSlot(other, CALENDAR);
-    const calendarId = ES.ToString(calendar);
-    const otherCalendarId = ES.ToString(otherCalendar);
-    if (calendarId !== otherCalendarId) {
-      throw new RangeError(`cannot compute difference between dates of ${calendarId} and ${otherCalendarId} calendars`);
-    }
-    options = ES.GetOptionsObject(options);
-    const smallestUnit = ES.ToSmallestTemporalUnit(options, 'nanosecond');
-    const defaultLargestUnit = ES.LargerOfTwoTemporalUnits('hour', smallestUnit);
-    const largestUnit = ES.ToLargestTemporalUnit(options, 'auto', [], defaultLargestUnit);
-    ES.ValidateTemporalUnitRange(largestUnit, smallestUnit);
-    const roundingMode = ES.ToTemporalRoundingMode(options, 'trunc');
-    const roundingIncrement = ES.ToTemporalDateTimeRoundingIncrement(options, smallestUnit);
-
-    const ns1 = GetSlot(this, EPOCHNANOSECONDS);
-    const ns2 = GetSlot(other, EPOCHNANOSECONDS);
-    let years, months, weeks, days, hours, minutes, seconds, milliseconds, microseconds, nanoseconds;
-    if (largestUnit !== 'year' && largestUnit !== 'month' && largestUnit !== 'week' && largestUnit !== 'day') {
-      // The user is only asking for a time difference, so return difference of instants.
-      years = 0;
-      months = 0;
-      weeks = 0;
-      days = 0;
-      ({ seconds, milliseconds, microseconds, nanoseconds } = ES.DifferenceInstant(
-        ns1,
-        ns2,
-        roundingIncrement,
-        smallestUnit,
-        roundingMode
-      ));
-      ({ hours, minutes, seconds, milliseconds, microseconds, nanoseconds } = ES.BalanceDuration(
-        0,
-        0,
-        0,
-        seconds,
-        milliseconds,
-        microseconds,
-        nanoseconds,
-        largestUnit
-      ));
-    } else {
-      const timeZone = GetSlot(this, TIME_ZONE);
-      if (!ES.TimeZoneEquals(timeZone, GetSlot(other, TIME_ZONE))) {
-        throw new RangeError(
-          "When calculating difference between time zones, largestUnit must be 'hours' " +
-            'or smaller because day lengths can vary between time zones due to DST or time zone offset changes.'
-        );
-      }
-      const untilOptions = { ...options, largestUnit };
-      ({ years, months, weeks, days, hours, minutes, seconds, milliseconds, microseconds, nanoseconds } =
-        ES.DifferenceZonedDateTime(ns1, ns2, timeZone, calendar, largestUnit, untilOptions));
-      ({ years, months, weeks, days, hours, minutes, seconds, milliseconds, microseconds, nanoseconds } =
-        ES.RoundDuration(
-          years,
-          months,
-          weeks,
-          days,
-          hours,
-          minutes,
-          seconds,
-          milliseconds,
-          microseconds,
-          nanoseconds,
-          roundingIncrement,
-          smallestUnit,
-          roundingMode,
-          this
-        ));
-      ({ years, months, weeks, days, hours, minutes, seconds, milliseconds, microseconds, nanoseconds } =
-        ES.AdjustRoundedDurationDays(
-          years,
-          months,
-          weeks,
-          days,
-          hours,
-          minutes,
-          seconds,
-          milliseconds,
-          microseconds,
-          nanoseconds,
-          roundingIncrement,
-          smallestUnit,
-          roundingMode,
-          this
-        ));
-    }
-
-    const Duration = GetIntrinsic('%Temporal.Duration%');
-    return new Duration(years, months, weeks, days, hours, minutes, seconds, milliseconds, microseconds, nanoseconds);
+    return ES.DifferenceTemporalZonedDateTime('until', this, other, options);
   }
   since(other, options = undefined) {
     if (!ES.IsTemporalZonedDateTime(this)) throw new TypeError('invalid receiver');
-    other = ES.ToTemporalZonedDateTime(other);
-    const calendar = GetSlot(this, CALENDAR);
-    const otherCalendar = GetSlot(other, CALENDAR);
-    const calendarId = ES.ToString(calendar);
-    const otherCalendarId = ES.ToString(otherCalendar);
-    if (calendarId !== otherCalendarId) {
-      throw new RangeError(`cannot compute difference between dates of ${calendarId} and ${otherCalendarId} calendars`);
-    }
-    options = ES.GetOptionsObject(options);
-    const smallestUnit = ES.ToSmallestTemporalUnit(options, 'nanosecond');
-    const defaultLargestUnit = ES.LargerOfTwoTemporalUnits('hour', smallestUnit);
-    const largestUnit = ES.ToLargestTemporalUnit(options, 'auto', [], defaultLargestUnit);
-    ES.ValidateTemporalUnitRange(largestUnit, smallestUnit);
-    let roundingMode = ES.ToTemporalRoundingMode(options, 'trunc');
-    roundingMode = ES.NegateTemporalRoundingMode(roundingMode);
-    const roundingIncrement = ES.ToTemporalDateTimeRoundingIncrement(options, smallestUnit);
-
-    const ns1 = GetSlot(this, EPOCHNANOSECONDS);
-    const ns2 = GetSlot(other, EPOCHNANOSECONDS);
-    let years, months, weeks, days, hours, minutes, seconds, milliseconds, microseconds, nanoseconds;
-    if (largestUnit !== 'year' && largestUnit !== 'month' && largestUnit !== 'week' && largestUnit !== 'day') {
-      // The user is only asking for a time difference, so return difference of instants.
-      years = 0;
-      months = 0;
-      weeks = 0;
-      days = 0;
-      ({ seconds, milliseconds, microseconds, nanoseconds } = ES.DifferenceInstant(
-        ns1,
-        ns2,
-        roundingIncrement,
-        smallestUnit,
-        roundingMode
-      ));
-      ({ hours, minutes, seconds, milliseconds, microseconds, nanoseconds } = ES.BalanceDuration(
-        0,
-        0,
-        0,
-        seconds,
-        milliseconds,
-        microseconds,
-        nanoseconds,
-        largestUnit
-      ));
-    } else {
-      const timeZone = GetSlot(this, TIME_ZONE);
-      if (!ES.TimeZoneEquals(timeZone, GetSlot(other, TIME_ZONE))) {
-        throw new RangeError(
-          "When calculating difference between time zones, largestUnit must be 'hours' " +
-            'or smaller because day lengths can vary between time zones due to DST or time zone offset changes.'
-        );
-      }
-      const untilOptions = { ...options, largestUnit };
-      ({ years, months, weeks, days, hours, minutes, seconds, milliseconds, microseconds, nanoseconds } =
-        ES.DifferenceZonedDateTime(ns1, ns2, timeZone, calendar, largestUnit, untilOptions));
-      ({ years, months, weeks, days, hours, minutes, seconds, milliseconds, microseconds, nanoseconds } =
-        ES.RoundDuration(
-          years,
-          months,
-          weeks,
-          days,
-          hours,
-          minutes,
-          seconds,
-          milliseconds,
-          microseconds,
-          nanoseconds,
-          roundingIncrement,
-          smallestUnit,
-          roundingMode,
-          this
-        ));
-      ({ years, months, weeks, days, hours, minutes, seconds, milliseconds, microseconds, nanoseconds } =
-        ES.AdjustRoundedDurationDays(
-          years,
-          months,
-          weeks,
-          days,
-          hours,
-          minutes,
-          seconds,
-          milliseconds,
-          microseconds,
-          nanoseconds,
-          roundingIncrement,
-          smallestUnit,
-          roundingMode,
-          this
-        ));
-    }
-
-    const Duration = GetIntrinsic('%Temporal.Duration%');
-    return new Duration(
-      -years,
-      -months,
-      -weeks,
-      -days,
-      -hours,
-      -minutes,
-      -seconds,
-      -milliseconds,
-      -microseconds,
-      -nanoseconds
-    );
+    return ES.DifferenceTemporalZonedDateTime('since', this, other, options);
   }
   round(roundTo) {
     if (!ES.IsTemporalZonedDateTime(this)) throw new TypeError('invalid receiver');
